% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Select_variables.R
\name{\%s\%}
\alias{\%s\%}
\title{Operator for Selecting Columns}
\usage{
dtf_data \%s\% chr_select
}
\arguments{
\item{dtf_data}{A data frame.}

\item{chr_select}{A character vector, the
specific variants of \code{swaap_select}
to call. Options currently include:
\code{'base'}, \code{'contact'},
\code{'demographics'}, \code{'experience'},
\code{'inventories'}, \code{'linked'},
\code{'linking'}, \code{'quality'},
\code{'SBIRT'}, and \code{'substances'}.}
}
\value{
A data frame with the subset of
specified columns.
}
\description{
Operator for calling \link{swaap_select}
and specified column subsets.
}
