% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swaap-package.R
\docType{package}
\name{swaap-package}
\alias{swaap-package}
\alias{swaap}
\title{School-Wide Assessment Analysis Preparation}
\description{
The \pkg{swaap} package provides common
functions for prepping school-wide assessment
data collected by the Center for School
Behavioral Health Massachusetts General
Hospital. In particular, the package
provides the linking code used to
connect records across time points.
}
\author{
\strong{Maintainer}: Kevin Potter \email{kpotter5@mgh.harvard.edu} (\href{https://orcid.org/0000-0002-3845-0512}{ORCID})

Other contributors:
\itemize{
  \item Michael Pascale [contributor]
  \item Jasmeen Kaur \email{jkaur12@mgh.harvard.edu} [contributor]
}

}
