% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R02-Add_variables.R
\name{swaap_add.ID_column}
\alias{swaap_add.ID_column}
\title{Add Identifier Variable}
\usage{
swaap_add.ID_column(dtf_data, chr_type)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}

\item{chr_label}{A character string, the
label for the identifier column, so
that a new column \code{'IDN.CHR.<Label>.ID'}
is created.}
}
\value{
A data frame with an additional
column.
}
\description{
Function to add an identifier variable
that includes information on the
year, semester, and grade level for
a record.
}
\author{
Kevin Potter
}
