% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R02-Add_variables.R
\name{swaap_add.quality}
\alias{swaap_add.quality}
\title{Add Variables for Quality Checks}
\usage{
swaap_add.quality(dtf_data)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}
}
\value{
A data frame with additional
columns
\code{'QLT.LGC.Remove'} and
\code{'QLT.CHR.Remove'} indicating
records that failed to meet various
quality standards.
}
\description{
Function to add variables with quality
check information. Includes a variable
flagging records to remove because
(a) they are missing a school code,
(b) they are missing a grade level,
(c) they are in a grade outside of 6-12.
}
\author{
Kevin Potter
}
