% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R02-Add_variables.R
\name{swaap_add.SBIRT_sample}
\alias{swaap_add.SBIRT_sample}
\title{Add Variables Tracking SBIRT Sample}
\usage{
swaap_add.SBIRT_sample(dtf_data)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}
}
\value{
A data frame with additional
columns \code{'SSS.LGC.SBIRT'} (\code{TRUE}
for students part of the SBIRT sample),
\code{'SSS.INT.SBIRTTimePoint'} (ranging
from 0 - 4, where 0 is the baseline time
point and 4 is month 24 of the study),
and \code{'SSS.INT.RecruitmentWave'}
(ranging from 1 - 3).
}
\description{
Function to add variables indicating the
subset of the data belonging to the
screening, brief intervention, and referral
to treatment (SBIRT) sample for the
school-wide assessment data, along with
variables tracking the time point.
}
\author{
Jasmeen Kaur; Kevin Potter
}
