% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R02-Add_variables.R
\name{swaap_add.school_enrollment}
\alias{swaap_add.school_enrollment}
\title{Add School Enrollment by Grade}
\usage{
swaap_add.school_enrollment(dtf_data)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}
}
\value{
A data frame with additional
columns \code{'SSS.INT.SchoolEnrollment'}
(the number of students enrolled for
the given school and grade), and
\code{'QLT.LGC.GradeSettoNA'} (when a
student reported a grade not in their
listed school).
}
\description{
Function that adds the number of students
enrolled for the given grade and school
for each record in the data set.
}
\author{
Kevin Potter
}
