% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R02-Add_variables.R
\name{swaap_add.source}
\alias{swaap_add.source}
\title{Add Details on Sources for Data}
\usage{
swaap_add.source(dtf_data, chr_source_files = "")
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}

\item{chr_source_files}{A character vector,
the standardized file names for the
source school-wide assessment data.}
}
\value{
A data frame with the additional
column \code{'SSS.CHR.SourceFileDetails'}.
Use the \code{attribute} function to
extract the details.
}
\description{
Function to add a column with the
attribute \code{'swaap.source_file_details'}
detailing the date and package version
used to prep data (and if supplied, the
source files for the school-wide assessment
data).
}
\author{
Kevin Potter
}
