% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R02-Add_variables.R
\name{swaap_add.time_point}
\alias{swaap_add.time_point}
\title{Add Variables for Time Points}
\usage{
swaap_add.time_point(dtf_data, lgc_grade = FALSE)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}

\item{lgc_grade}{A logical value; if
\code{TRUE} defines time points using grade
levels.}
}
\value{
A data frame with additional
columns \code{'SSS.INT.TimePoint'}
(starting from 0) and
\code{'SSS.INT.LongitudinalWave'}.
}
\description{
Function to add variables indicating the
time point, defined either by the
survey year and semester, or by grade
level. If by year and semester, clusters
by grades assuming standard progression
from one grade to next. If data consists
of the SBIRT sample only, uses the
pre-defined SBIRT time points and recruitment
wave.
}
\author{
Kevin Potter
}
