% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R02-Add_variables.R
\name{swaap_add.year_and_semester}
\alias{swaap_add.year_and_semester}
\title{Add Variables for School Year and Semester}
\usage{
swaap_add.year_and_semester(dtf_data)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}
}
\value{
A data frame with additional
columns
\code{'SSS.CHR.DataSet'},
\code{'SSS.CHR.SurveyYearSemester'},
\code{'SSS.CHR.TimePoint'}, and
\code{'SSS.CHR.Semester'} indicating
the survey year and semester.
}
\description{
Function to add variables indicating the
school year and semester for the
given school-wide assessment data.
}
\author{
Kevin Potter
}
