% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R07-Analysis_functions.R
\name{swaap_analysis.prep_project}
\alias{swaap_analysis.prep_project}
\title{Prep Analysis Project}
\usage{
swaap_analysis.prep_project(chr_name, chr_email, chr_project)
}
\arguments{
\item{chr_name}{The user's name.}

\item{chr_email}{The user's email.}

\item{chr_project}{The project label to begin
each function with.}
}
\value{
As a side effect creates the 'Source'
and 'R' folders, and creates a '_targets.R' and
'R01-Data_processing.R' scripts.
}
\description{
Function to generate folders and R script
templates for an analysis project looking
at the school-wide assessment data.
}
\author{
Kevin Potter
}
