% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R07-Analysis_functions.R
\name{swaap_analysis.schools}
\alias{swaap_analysis.schools}
\title{Create School-Level Summary}
\usage{
swaap_analysis.schools(dtf_data, chr_measure = NULL, fun_summary = NULL)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}

\item{chr_measure}{An optional character vector, the
additional columns in \code{dtf_data} over which to
create school-level summaries.}

\item{fun_summary}{A function to compute the univariate
statistic for the school-level summaries.}
}
\value{
A data frame.
}
\description{
Create a school-level data set and optionally create
school-level summaries for a specified set of measures.
}
