% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R07-Analysis_functions.R
\name{swaap_analysis.source_scripts}
\alias{swaap_analysis.source_scripts}
\title{Source in R scripts}
\usage{
swaap_analysis.source_scripts(chr_folder = "R")
}
\arguments{
\item{chr_folder}{A character string, an optional
path to the folder with the R scripts - by
default assumed to be \code{'R'}.}
}
\value{
As a side effect sources all .R files.
}
\description{
Function to source in code defined in a set of
.R files all at once.
}
\author{
Kevin Potter
}
