% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Data_functions.R
\name{swaap_data.at}
\alias{swaap_data.at}
\title{Take Slice of Values at Specified Time}
\usage{
swaap_data.at(
  dtf_data,
  int_time,
  chr_column,
  chr_ID = "IDN.CHR.Linked.ID",
  chr_new = "",
  obj_default = NA
)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
have linked records across time points
and a column \code{'SSS.INT.TimePoint'}.}

\item{int_time}{An integer value, the time
point of interest (must be a value in
\code{'SSS.INT.TimePoint'}),}

\item{chr_column}{A character string, the
column name with the values to subset.}

\item{chr_ID}{A character string, the column
name for the linked record identifier.}

\item{chr_new}{A character string, the new
column name to add to \code{dtf_data}
with the propagated values. If blank
instead returns a new data frame.}

\item{obj_default}{A value, the default to
return when the time point is missing.}
}
\value{
A data frame, either the extracted
values at the specified time for each
ID or the updated original data frame.
}
\description{
Function that will extract values for a
variable at a specified time point (and
if a new variable is given, update the
data frame with those values propagated
by ID).
}
