% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Data_functions.R
\name{swaap_data.attr}
\alias{swaap_data.attr}
\title{Extract Attributes Specific to swaap package}
\usage{
swaap_data.attr(vec_values)
}
\arguments{
\item{vec_values}{A vector of values with
attributes.}
}
\value{
The \code{swaap}-created attribute.
}
\description{
Function to extract attribute of a column
created as part of a \code{swaap} function
call.
}
