% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Data_functions.R
\name{swaap_data.codebook_names}
\alias{swaap_data.codebook_names}
\title{Match Variables Names to Codebook}
\usage{
swaap_data.codebook_names(
  dtf_data = NULL,
  chr_columns = NULL,
  lgc_reverse = FALSE
)
}
\arguments{
\item{dtf_data}{A data frame.}

\item{chr_columns}{A character vector.}

\item{lgc_reverse}{A logical value; if \code{TRUE}
converts back to original variable names instead.}
}
\value{
Either a data frame or character vector,
depending if \code{dtf_data} or \code{chr_columns}
is provided, respectively. If no input is given,
returns a data frame with the pairs of
original and codebook variable names.
}
\description{
Convenience function that will convert
variable names to be consistent with
the codebook for shared linked SURF data.
}
