% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Data_functions.R
\name{swaap_data.deidentified}
\alias{swaap_data.deidentified}
\title{Deidentify a Standardized Data Set}
\usage{
swaap_data.deidentified(dtf_data, lgc_codebook_names = TRUE)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}
}
\value{
A data frame with columns containing
potential identifying data removed.
}
\description{
Function to deidentify data following
the standardized format for school-wide
assessment data.
}
