% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Data_functions.R
\name{swaap_data.files}
\alias{swaap_data.files}
\title{Path to School-wide Assessment Data Files}
\usage{
swaap_data.files(
  chr_dir,
  chr_tag = "SWA-Surveys-",
  chr_sep = "-",
  lgc_full = TRUE
)
}
\arguments{
\item{chr_dir}{The path to the folder with the
files to consider.}

\item{chr_tag}{A character string, the standardized
tag used to indicate data files for the
school-wide assessment.}

\item{chr_sep}{A character string, the separator
between the label and subsequent file info.'}

\item{lgc_full}{A logical value; if \code{TRUE}
returns the full conditional path of the files.}
}
\value{
A labeled character vector of file paths.
}
\description{
Given a standardized starting label, identifies
and returns the path to data files for the
school-wide assessment survey data. Files are
assumed to start with a leading tag, followed
by a label (typically the year and semester),
followed by a separator and additional information
(e.g., the date).
}
\author{
Kevin Potter
}
