% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Data_functions.R
\name{swaap_data.merge}
\alias{swaap_data.merge}
\title{Merge Multiple Data Sets}
\usage{
swaap_data.merge(
  lst_data,
  chr_add = "",
  chr_recode = "",
  chr_select = "",
  lgc_SBIRT = FALSE,
  lgc_remove_flagged = TRUE,
  lgc_progress = TRUE
)
}
\arguments{
\item{lst_data}{A list of data frames, assumed
to be standard school-wide assessment survey
data. Can also be \code{NULL}, in which case
templates for the \code{chr_add}, \code{chr_recode},
and \code{chr_select} variables are provided.}

\item{chr_add}{A character vector, the \code{swaap_add}
functions to run. Either takes the full function
name, or as a shorthand the part of the function
name following the period (e.g., instead of
\code{'swaap_add.school_enrollment'} you can just
use \code{'school_enrollment'}).}

\item{chr_recode}{A character vector, the \code{swaap_recode}
functions to run. Either takes the full function
name, or as a shorthand the part of the function
name following the period (e.g., instead of
\code{'swaap_recode.base'} you can just
use \code{'base'}).}

\item{chr_select}{A character vector, the \code{swaap_select}
functions to run. Either takes the full function
name, or as a shorthand the part of the function
name following the period (e.g., instead of
\code{'swaap_select.base'} you can just
use \code{'base'}).}

\item{lgc_SBIRT}{A logical value; if \code{TRUE}
restricts the output only to the SBIRT sample.}

\item{lgc_remove_flagged}{A logical value; if \code{TRUE}
remove records that fail initial quality checks
(see \link{swaap_add.quality_checks}).}

\item{lgc_progress}{A logical value; if \code{TRUE}
displays the function's progress.}
}
\value{
A data frame.
}
\description{
Functions to merge multiple school-wide
assessment data sets into single long-form
data set.
}
