% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Data_functions.R
\name{swaap_data.missing}
\alias{swaap_data.missing}
\title{Tally NA Values Over Columns}
\usage{
swaap_data.missing(
  dtf_data,
  chr_columns = NULL,
  chr_new = "",
  chr_output = "count"
)
}
\arguments{
\item{dtf_data}{A data frame.}

\item{chr_columns}{An optional character vector,
the subset of columns in \code{dtf_data} to
consider. If not provided, all columns are
considered.}

\item{chr_new}{A character string,
the column name to add to \code{dtf_data}
with the desired output. If \code{''}
(default) no column is added.}

\item{chr_output}{A character string, the
type of output to return. If \code{'count'}
returns the total number of NA values per
row; if \code{'proportion'} returns the
proportion of NA values per row; if
\code{'pattern'} returns a character
string of 0s and 1s per row where
1s denote NA values.}
}
\value{
Either a vector or a data frame.
}
\description{
Function that tallies the number
of \code{NA} values per row over
columns in a data set.
}
