% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Data_functions.R
\name{swaap_data}
\alias{swaap_data}
\title{Generate data}
\usage{
swaap_data(chr_type, ...)
}
\arguments{
\item{chr_type}{A character vector of up to
two elements, the type and sub-type of
data to generate.}

\item{...}{Additional arguments to pass to the
internal data generation function.}
}
\value{
A data frame.
}
\description{
Function to generate assorted data relevant to
the school-wide assessment project.
}
\examples{
dtf_debug <- swaap_data( c('link', 'debug') )

}
