% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Data_functions.R
\name{swaap_data.replace}
\alias{swaap_data.replace}
\title{Replace Values per User Specification}
\usage{
swaap_data.replace(
  vec_values,
  lst_comparison,
  chr_action,
  vec_replacement,
  obj_default
)
}
\arguments{
\item{vec_values}{A vector of values.}

\item{lst_comparison}{A list of of vectors, different
comparisons to test against \code{vec_values}.}

\item{chr_action}{A character vector, the type of
action to conduct for each comparison given in
\code{lst_comparison}.}

\item{vec_replacement}{A vector of values matching
in length with \code{lst_comparison} and \code{chr_action},
what to return when a given condition is met.}

\item{obj_default}{A value, the default output to return
if no conditions are met.}
}
\value{
A single value as defined in \code{vec_replacement}.
}
\description{
Function to replace a vector of values with a
single value based on conditional logic.
}
\examples{
swaap_data.replace( c( 'A', NA ), list( 'A' ), 'all', 'A', NA )

}
\author{
Kevin Potter
}
