% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Data_functions.R
\name{swaap_data.replace_defaults}
\alias{swaap_data.replace_defaults}
\title{Default Options for swaap_data.replace}
\usage{
swaap_data.replace_defaults(chr_column, chr_option = "", chr_coding = "")
}
\arguments{
\item{chr_column}{A character string, the column
name.}

\item{chr_option}{A character string, additional
options to consider when generating inputs.}

\item{chr_coding}{A character string. Use the
options \code{'dummy'} or \code{'effect'} to
convert values to a desired numeric coding scheme.}
}
\value{
A list of inputs for \link{swaap_data.replace}.
}
\description{
Function that generates a list of inputs for
\link{swaap_data.replace} based on a given
column and additional options.
}
\author{
Kevin Potter
}
