% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Data_functions.R
\name{swaap_data.static}
\alias{swaap_data.static}
\title{Create Static Variant of a Variable}
\usage{
swaap_data.static(
  dtf_data,
  chr_variable,
  chr_new = "",
  chr_ID = "IDN.CHR.Linked.ID",
  lst_replace = NULL
)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}

\item{chr_variable}{A character string, the
column to update to ensure static values
across time points.}

\item{chr_new}{A character vector of up to
two values, the new column names for
(1) the revised static variable, and
(2) a quality check variable indicating
cases that had to be recoded to be static.}

\item{chr_ID}{A character string, the column
name for the identifier.}

\item{lst_replace}{An optional list of
inputs for \link{swaap_data.replace}.}
}
\value{
A data frame with up to two new columns.
}
\description{
Function that will create a static variable
in which values are consistent across multiple
time points.
}
