% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Data_functions.R
\name{swaap_data.subset}
\alias{swaap_data.subset}
\title{Subset Data Frame While Perserving Attributes}
\usage{
swaap_data.subset(dtf_data, vec_subset)
}
\arguments{
\item{dtf_data}{A data frame.}

\item{vec_subset}{A vector, either a logical vector
or a vector of indices specifying a subset of
\code{dtf_data} to take.}
}
\value{
A data frame.
}
\description{
Function to subset a data frame while preserving
any column attributes.
}
\author{
Kevin Potter
}
