% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Data_functions.R
\name{swaap_data.survey_summary}
\alias{swaap_data.survey_summary}
\title{Summary of Survey Details}
\usage{
swaap_data.survey_summary(dtf_SRV, chr_times = NULL)
}
\arguments{
\item{dtf_SRV}{A data frame, assumed to be
output from \link{swaap_data.merge}.}

\item{chr_times}{A character vector with
elements in the form \code{'YYYY Fall'}
or \code{'YYYY Spring'}, indicating which
time points to include.}
}
\value{
Output to the console window with
relevant details.
}
\description{
Function that provides summary of
survey (number of surveys, number removed
due to data cleaning, breakdowns by
school type, etc.).
}
