% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Linking_code.R
\name{swaap_link.helper.input}
\alias{swaap_link.helper.input}
\title{Function to Help Create Input for Linking Function}
\usage{
swaap_link.helper.input(
  dtf_long,
  chr_input,
  lst_sets = NULL,
  obj_extra = NULL,
  lgc_progress = FALSE
)
}
\arguments{
\item{dtf_long}{A data frame, assumed to be standard processed
school-wide assessment data with the columns
\code{'SSS.INT.TimePoint'} and \code{'SSS.INT.LongitudinalWave'}
as well as the linking code items.}

\item{chr_input}{A character string indicating the type of input to
create, either \code{'sets'}, \code{'items'}, \code{'combos'},
or \code{'missing'}.}

\item{lst_sets}{An optional argument, include to aid in defining
items. A list of lists, with each sublist specifying
\code{'Base'} and \code{'Add'} logical vectors for the pair of data
subsets in \code{dtf_long} to link over (e.g., \code{'Base'} would
subset the first time point and \code{'Add'} would subset the second
time point). Include to aid in defining items, combos, and
missingness treatment.}

\item{obj_extra}{Additional input needed to define combos and
treatment of missingness. For combos, either a character vector
or list of the items; for missingness, a list with both the
items and combos input.}

\item{lgc_progress}{A logical value; if\code{TRUE} tracks function
progress.}
}
\value{
A list structured in the relevant way for the given output.
}
\description{
Helper function to generate input to pass to the \link{swaap_link}
function.
}
\examples{
dtf_long <- swaap_simulate( 'link', 'demo' )

lst_sets <- dtf_long |>
  swaap_link.helper.input( 'sets' )
lst_items <- dtf_long |>
  swaap_link.helper.input( 'items', lst_sets = lst_sets )
lst_combos <- dtf_long |>
  swaap_link.helper.input( 'combos', obj_extra = lst_items )
lst_missing <- dtf_long |>
  swaap_link.helper.input(
    'missing', obj_extra = list( lst_items, lst_combos )
  )

}
