% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Linking_code.R
\name{swaap_link.helper.rows}
\alias{swaap_link.helper.rows}
\title{Extract Rows Flagged for Linking}
\usage{
swaap_link.helper.rows(dtf_linked, int_row = 1)
}
\arguments{
\item{dtf_linked}{A data frame, output from the
\link{swaap_link} function. Must have the column
\code{LNK.CHR.Rows}.}

\item{int_row}{An integer, the row of \code{dtf_linked}
to consider.}
}
\value{
Either the subset of rows in \code{dtf_linked}
that were flagged for linking, or (if \code{dtf_linked}
consists of only a single row) an integer vector.
}
\description{
Function to extract the rows flagged for linking after
running \link{swaap_link}.
}
\author{
Kevin Potter
}
