% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Linking_code.R
\name{swaap_link.helper.similarity}
\alias{swaap_link.helper.similarity}
\title{Compute Post-hoc Similarity Score for Linked Records}
\usage{
swaap_link.helper.similarity(dtf_linked, chr_items, int_times = NULL)
}
\arguments{
\item{dtf_linked}{A data frame, output from the
\link{swaap_link} function. Must have the columns
\code{IDN.CHR.Linked.ID}, \code{LNK.LGC.NoIssues},
and \code{IDN.CHR.Linked.ID}.}

\item{chr_items}{A character vector, the columns to
compare when computing the similarity score.}

\item{int_times}{An integer vector, the pair of
of time points to compute the score over.}
}
\value{
A list with the matrix of matches by linked ID,
a data frame summarizing similarity by item, and a
data frame with similarity scores by ID.
}
\description{
Function to compute a post-hoc similarity score based
on exact matches for records across a pair of time
points.
}
\author{
Kevin Potter
}
