% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Linking_code.R
\name{swaap_link.helper.trim}
\alias{swaap_link.helper.trim}
\title{Flag Duplicate Records to Trim Based on Rules}
\usage{
swaap_link.helper.trim(dtf_long, fun_rule = NULL, lst_arg = NULL)
}
\arguments{
\item{dtf_long}{A data frame, assumed to be standard processed
school-wide assessment data with the columns
\code{'LNK.LGC.Duplicates'} and \code{'IDN.CHR.Linked.ID'}.}

\item{fun_rule}{A function that takes a data frame
and returns a logical vector equal to the number
of rows set to \code{TRUE} for rows to keep and
\code{FALSE} otherwise.}

\item{lst_arg}{An optional list of additional arguments to
pass to the function for trimming.}
}
\value{
A data frame with a column \code{'QLT.LGC.RemoveDuplicate'}
indicating rows to remove that fail to meet the rules.
}
\description{
Function to flag duplicate records based
on automated rules (default is to select
first record with highest rate of completion)
for subsequent trimming via the column
\code{'QLT.LGC.RemoveDuplicate'}.
}
\author{
Kevin Potter
}
