% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Linking_code.R
\name{swaap_link.input.combos}
\alias{swaap_link.input.combos}
\title{Function to Define Combinations of Items to Link Over}
\usage{
swaap_link.input.combos(
  dtf_long,
  lst_items,
  lst_combos = NULL,
  lgc_fastLink = FALSE
)
}
\arguments{
\item{dtf_long}{A data frame, assumed to be standard processed
school-wide assessment data.}

\item{lst_items}{A list of of character vectors, the columns to
use as linking items for each set (the pair of time points to link).
See output from \link{swaap_link.input.items}.}

\item{lst_combos}{An optional argument. For exact matching methods,
should be a list of lists, each sublist consisting of integer
indices indicating the combination of linking items to use.
For the \link[fastLink:fastLink]{fastLink::fastLink} function, a list of lists, each
sublist specifying integer indices indicating which items to
pass to the \code{stringdist}, \code{numeric},
and \code{partial} arguments of the function \link[fastLink:fastLink]{fastLink::fastLink}.
If provided, function checks if input is valid.}

\item{lgc_fastLink}{A logical value; if \code{TRUE} specifies
which items to pass to the \code{stringdist}, \code{numeric},
and \code{partial} arguments of the function \link[fastLink:fastLink]{fastLink::fastLink}
when generating input.}
}
\value{
A list of lists. For exact matching methods, each sublist
consists of integer indices indicating the combination of linking
items to use. For the \link[fastLink:fastLink]{fastLink::fastLink} function, each sublist
specifies the integer indices to pass to the \code{stringdist},
\code{numeric}, and \code{partial} arguments.
}
\description{
Function to generate input to pass to the \link{swaap_link}
function. Creates a list of lists, either the integer indices
for the combinations of linking items to consider for
exact matching approaches, or the integer indices for the
items to pass to pass to the \code{stringdist}, \code{numeric},
and \code{partial} arguments of the function \link[fastLink:fastLink]{fastLink::fastLink}.
}
\examples{
dtf_long <- swaap_simulate( 'link', 'demo' )

lst_sets <- dtf_long |>
  swaap_link.input.sets()
lst_items <- dtf_long |>
  swaap_link.input.items( lst_sets )
lst_combos <- dtf_long |>
  swaap_link.input.combos( lst_items )

}
