% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Linking_code.R
\name{swaap_link.input.items}
\alias{swaap_link.input.items}
\title{Function to Define Items to Link Over}
\usage{
swaap_link.input.items(
  dtf_long,
  lst_sets,
  obj_items = NULL,
  lgc_fastLink = FALSE,
  lgc_district = FALSE
)
}
\arguments{
\item{dtf_long}{A data frame, assumed to be standard processed
school-wide assessment data. If column \code{'SSS.INT.Grade'}
is present function will try to check for instances in which
school code does not apply (i.e., transition from middle to
high school).}

\item{lst_sets}{A list of lists, each sublist consisting of the
elements \code{Base} (a logical vector for rows from the first
time point) and \code{Add} (a logical vector for rows from the
second time point). See output from \link{swaap_link.input.sets}.}

\item{obj_items}{An optional argument, either a character vector
of column names or a list of character vectors. If a character
vector, function converts into a list using \code{lst_sets}.
If a list, function checks if input is valid.}

\item{lgc_fastLink}{A logical value; if \code{TRUE}, will
return linking items intended for use with the
\link[fastLink:fastLink]{fastLink::fastLink} function when generating input.}

\item{lgc_district}{A logical value; if \code{TRUE} uses
district codes instead of school codes when generating input.}
}
\value{
A list of character vectors, the column names for the
items to link over per each pair of time points.
}
\description{
Function to generate input to pass to the \link{swaap_link}
function. Creates a list of the column names for the
items to link over per each pair of time points.
}
\examples{
dtf_long <- swaap_simulate( 'link', 'demo' )

lst_sets <- dtf_long |>
  swaap_link.input.sets()
lst_items <- dtf_long |>
  swaap_link.input.items( lst_sets )

}
