% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Linking_code.R
\name{swaap_link.input.missing}
\alias{swaap_link.input.missing}
\title{Function to Define Missing Items to Ignore}
\usage{
swaap_link.input.missing(dtf_long, lst_items, lst_combos, lst_missing = NULL)
}
\arguments{
\item{dtf_long}{A data frame, assumed to be standard processed
school-wide assessment data.}

\item{lst_items}{A list of of character vectors, the columns to
use as linking items for each set (the pair of time points to link).
See output from \link{swaap_link.input.items}.}

\item{lst_combos}{A list of lists, each sublist consisting of integer
indices indicating the combination of linking items to use.
See output from \link{swaap_link.input.combos}.}

\item{lst_missing}{An optional argument, a list of lists, each
sublist consisting of integer indices for items that must be
non-missing for linking to be attempted. If provided,
function checks if input is valid.}
}
\value{
A list of lists, each sublist consisting of integer indices
for items that must be non-missing for linking to be attempted.
}
\description{
Function to generate input to pass to the \link{swaap_link}
function. Creates a list of lists, the integer indices
for the linking items that must be non-missing for linking
to be attempted with exact matching methods.
}
\examples{
dtf_long <- swaap_simulate( 'link', 'demo' )

lst_sets <- dtf_long |>
  swaap_link.input.sets()
lst_items <- dtf_long |>
  swaap_link.input.items( lst_sets )
lst_combos <- dtf_long |>
  swaap_link.input.combos( lst_items )
lst_missing <- dtf_long |>
  swaap_link.input.missing( lst_items, lst_comobs )

}
