% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Linking_code.R
\name{swaap_link.input.sets}
\alias{swaap_link.input.sets}
\title{Function to Define Sets of Records to Link}
\usage{
swaap_link.input.sets(
  dtf_long,
  lst_sets = NULL,
  lgc_duplicates = FALSE,
  chr_groups = NULL
)
}
\arguments{
\item{dtf_long}{A data frame, assumed to be standard processed
school-wide assessment data. Default methods need the columns
\code{'SSS.INT.LongitudinalWave'}, \code{'SSS.INT.TimePoint'} and
\code{'SSS.INT.Grade'}.}

\item{lst_sets}{An optional argument, the list of sets. If provided,
function checks if input is valid.}

\item{lgc_duplicates}{A logical value; if \code{TRUE}, produces
the list of lists with logical vectors for checking for duplicate
records within a given time point.}

\item{chr_groups}{A character vector of three column names for the
longitudinal wave, time point, and grade level, respectively.}
}
\value{
A list of lists, each sublist consisting of the elements
\code{Base} (a logical vector for rows from the first time point)
and \code{Add} (a logical vector for rows from the second time point),
If \code{lgc_duplicates = TRUE}, both logical vectors are for the
same time point.
}
\description{
Function to generate input to pass to the \link{swaap_link}
function. Creates a list of lists with logical vectors
indicating the rows to consider when linking two time points.
Can also be used to check that input is correctly specified.
}
\examples{
dtf_long <- swaap_simulate( 'link', 'demo' )

lst_sets <- dtf_long |>
  swaap_link.input.sets()

}
\author{
Kevin Potter
}
