% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Linking_code.R
\name{swaap_link.parameters}
\alias{swaap_link.parameters}
\title{Extract Linkage Parameters}
\usage{
swaap_link.parameters(dtf_linked, chr_input = "lst_items", lgc_unlist = FALSE)
}
\arguments{
\item{dtf_linked}{A data frame, output from the
\link{swaap_link} function. Must have the column
\code{LNK.CHR.AttributeWithParameters}. Alternatively,
set to \code{NULL} to see a reminder of the possible
options for \code{'chr_input'}.}

\item{chr_input}{A character string, either \code{'chr_method'},
\code{'lst_sets'}, \code{'lst_items'}, \code{'lst_combos'},
or \code{'lst_missing'}.}

\item{lgc_unlist}{A logical value; if \code{TRUE} unlists
an object and returns the unique elements (useful for
the \code{'lst_items'} option).}
}
\value{
The specified parameter input used when linking records.
}
\description{
Function to extract the parameters used when
running \link{swaap_link} (e.g., \code{'lst_sets'},
\code{'lst_items'}, etc.).
}
\author{
Kevin Potter
}
