% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Linking_code.R
\name{swaap_link.report.comparison}
\alias{swaap_link.report.comparison}
\title{Compare Overlap Between Linked Records}
\usage{
swaap_link.report.comparison(dtf_linked, chr_base, chr_contrast)
}
\arguments{
\item{dtf_linked}{A data frame.}

\item{chr_base}{A character vector with 2 column names,
the linked IDs and time points for the base approach.}

\item{chr_contrast}{A character vector with 2 column names,
the linked IDs and time points for the contrasting approach.}
}
\value{
A data frame grouped by IDs from the base approach.
See the attributes of the \code{Summaries} column for
details on correct categorization from the contrasting approach.
}
\description{
Function, given a data set with linked IDs
and associated time points from two different
approaches, computes the subset of linked and
unique records from one approach categorized
similarly or differently from the other approach.
}
