% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Linking_code.R
\name{swaap_link.report.discrepant}
\alias{swaap_link.report.discrepant}
\title{Determine Discrepancies Between Records}
\usage{
swaap_link.report.discrepant(
  dtf_linked,
  chr_items,
  chr_missingness = "exclude",
  lgc_plot = TRUE,
  num_adj = c(0.05, 1.05),
  mat_layout = NULL,
  chr_colors = c("lightblue", "pink")
)
}
\arguments{
\item{dtf_linked}{A data frame, must have the
columns \code{'IDN.CHR.Linked.ID'} &
\code{'LNK.LGC.NoIssues'}.}

\item{chr_items}{A character vector, the items
to check for discrepancies across pairs of
records.}

\item{chr_missingness}{A character string,
used to specify the treatment of missing
values when comparing across records.
If \code{'match'} marks an item with
both record values as missing as stable,
and an item with only one value as
missing as discrepant. If \code{'ignore'}
marks an item as stable irrespective if
one or both values are missing. If
\code{'exclude'} removes the pair of
records from consideration if any
missing values are found.}

\item{lgc_plot}{A logical value; if \code{TRUE}
generates a plot summarizing the results.}

\item{num_adj}{...}

\item{mat_layout}{...}

\item{chr_colors}{...}
}
\value{
A list with a) a data frame tracking
stable and discrepant items per each pair of
linked records, b) a data frame tracking the
number of pairs that follow a specific pattern
of stable/discrepant values, and c) a data frame
tracking the marginal stability rates per each
item.
}
\description{
Given a data frame with linked records across
a pair of time points, computes discrepancies
between these records over a specified set of
items. Will generate a plot summarizing
findings if specified. Users should carefully
consider treatment of missing values when
}
