% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Linking_code.R
\name{swaap_link.report}
\alias{swaap_link.report}
\title{Summary of Linking of Records}
\usage{
swaap_link.report(
  dtf_linked,
  lst_groups = list(Wave = "SSS.INT.LongitudinalWave", Time = "SSS.INT.TimePoint")
)
}
\arguments{
\item{dtf_linked}{A data frame, the output
from \link{swaap_link}.}

\item{lst_groups}{A named list of column names,
the grouping factors to consider when summarizing
the number of records linked.}
}
\value{
A list of data frames.
}
\description{
Function to summarize the performance of
the linking code for the school-wide
assessment. Provides summary statistics
for what records were linked, and if
columns with the 'true' IDs are detected,
provides details on hits and correct
rejections.
}
\examples{
# Linking across time points
dtf_long <- swaap_simulate( 'link', 'demo' )
dtf_linked <- swaap_link(dtf_long)
lst_summary <- swaap_link.report(dtf_linked)

}
\author{
Kevin Potter
}
