% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Linking_code.R
\name{swaap_link.sets}
\alias{swaap_link.sets}
\title{Summary of Sets for Linking}
\usage{
swaap_link.sets(dtf_data, lst_sets, lgc_character = FALSE)
}
\arguments{
\item{dtf_data}{A data frame, assumed to be standard processed
school-wide assessment data with the columns
\code{'SSS.INT.SurveyYear'}, \code{'SSS.CHR.Semester'},
and \code{'SSS.INT.Grade'}.}

\item{lst_sets}{A list of lists, with each sublist specifying
\code{'Base'} and \code{'Add'} logical vectors for the pair of data
subsets in \code{dtf_long} to link over (e.g., \code{'Base'} would
subset the first time point and \code{'Add'} would subset the second
time point).}

\item{lgc_character}{A logical value; if \code{TRUE} outputs a
character vector instead of a data frame.}
}
\value{
Either a data frame with a row for each set of pairs, or
a character vector with concatenated abbreviations describing the
sets.
}
\description{
Returns a table summarizing the set of pairs to link
over in terms of year, semester, and grade level.
}
\author{
Kevin Potter
}
