% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Linking_code.R
\name{swaap_link.timepoints}
\alias{swaap_link.timepoints}
\title{Extract Possible Linked Time Point Patterns}
\usage{
swaap_link.timepoints(dtf_data, lgc_character = TRUE)
}
\arguments{
\item{dtf_data}{A data frame, assumed to be standard processed
school-wide assessment data with the columns
\code{'SSS.INT.TimePoint'} and \code{'LNK.CHR.TimePoints'}.}
}
\value{
Either a character vector with all possible linked
time point patterns (ignoring duplicates), or a logical
vector for all records that meet this linked time point
patterns.
}
\description{
Function to extract possible linked time point patterns
based on available time points, or isolate records that
meet linked time point patterns.
}
\author{
Kevin Potter
}
