% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Linking_code.R
\name{swaap_link.trim_rule}
\alias{swaap_link.trim_rule}
\title{Return Function for Rule to Trim Duplicates}
\usage{
swaap_link.trim_rule(chr_rule)
}
\arguments{
\item{chr_rule}{A character string, the type of rule
to use. Options include \code{'completed'} (choose
the record with the highest completion rate), or
\code{'outcome and completed'} (choose the record
that has a non-missing value for an outcome then
choose the record with the highest completion rate).}
}
\value{
A data frame with a column \code{'QLT.LGC.RemoveDuplicate'}
indicating rows to remove that fail to meet the rules.
}
\description{
Function to return another function that implements
a desired rule to trim flagged duplicates from a linked
data set.
}
\author{
Kevin Potter
}
