% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-Recode_variables.R
\name{swaap_recode.base}
\alias{swaap_recode.base}
\title{Recode Base Information Variables}
\usage{
swaap_recode.base(dtf_data)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}
}
\value{
A data frame with the additional variables
\code{'SSS.INT.DistrictCode'},
\code{'SSS.INT.SchoolCode'},
\code{'IDN.CHR.Record.ID'},
\code{'IDN.CHR.LocallyAssignedSchool.ID'}, and
\code{'IDX.INT.Origin.LASID'}.
}
\description{
Renames variables with base information
(school codes, record identifiers, etc.).
}
\author{
Kevin Potter
}
