% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-Recode_variables.R
\name{swaap_recode.character}
\alias{swaap_recode.character}
\title{Recode Character Value Columns to Integer}
\usage{
swaap_recode.character(dtf_data, lgc_mapping = FALSE)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}

\item{lgc_mapping}{A logical value; if
\code{TRUE} instead returns a data
frame with the mapping between
character and numeric values for
each changed variable.}
}
\value{
A data frame with variables with
the data type \code{CHR} (character strings)
renamed to \code{INT} (integer values), or
a data frame reporting on the mapping from
original character strings to numeric values.
}
\description{
Converts columns of character strings to
integers for compatibility with other
statistical software programs. Also
converts logical variables (\code{TRUE/FALSE})
to (\code{1/0}) instead.
}
\author{
Kevin Potter
}
