% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-Recode_variables.R
\name{swaap_recode.demographics}
\alias{swaap_recode.demographics}
\title{Recode Standard Demographic Variables}
\usage{
swaap_recode.demographics(dtf_data)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}
}
\value{
A data frame with the additional variables
\code{'SBJ.INT.AgeInYears'},
\code{'SBJ.CHR.Sex'},
\code{'SBJ.CHR.GenderIdentity'},
\code{'SBJ.CHR.Sexuality'},
\code{'SBJ.CHR.Race'}, and
\code{'SBJ.CHR.Ethnicity'}.
}
\description{
Renames standard demographic variables
(age, biological sex, etc.). Prioritizes
revised versions ending with \code{.R}.
}
