% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-Recode_variables.R
\name{swaap_recode.discrimination}
\alias{swaap_recode.discrimination}
\title{Recode Discrimination Items}
\usage{
swaap_recode.discrimination(dtf_data)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}
}
\value{
A data frame with the additional variables
\code{'INV.INT.DISC.Q1.Gender'},
\code{'INV.INT.DISC.Q2.Sexuality'},
\code{'INV.INT.DISC.Q3.Religion'},
\code{'INV.INT.DISC.Q4.Disability'},
\code{'INV.INT.DISC.Q5.Money'}, and
\code{'INV.INT.DISC.Q6.Other'}.
}
\description{
Recodes discrimination items.
}
\author{
Kevin Potter
}
