% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-Recode_variables.R
\name{swaap_recode.intermittent}
\alias{swaap_recode.intermittent}
\title{Recode Variables at Intermittent Time Points}
\usage{
swaap_recode.intermittent(dtf_data)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}
}
\value{
A data frame with the additional variables
\code{'SBJ.LGC.CloseConnection.<Type>'},
\code{'SBJ.INT.CloseConnection.Happiness'},
\code{'SBJ.LGC.FirstLanguage.English'} and
\code{'SBJ.LGC.HomeLanguage.English'}.
}
\description{
Renames variables that occurred at
intermittent time points.
}
\author{
Kevin Potter
}
