% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Select_variables.R
\name{swaap_select.base}
\alias{swaap_select.base}
\title{Select Base Columns}
\usage{
swaap_select.base(chr_input = NULL, lgc_original = FALSE)
}
\arguments{
\item{chr_input}{An optional character
vector, additional columns to include.}

\item{lgc_original}{A logical value; if
\code{TRUE} uses original variable
names instead of recoded ones.}
}
\value{
A character vector.
}
\description{
Function to select columns with basic
session information (e.g.,
survey year, grade level, etc.).
}
\author{
Kevin Potter
}
