% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Select_variables.R
\name{swaap_select.contact}
\alias{swaap_select.contact}
\title{Select Contact Information Items}
\usage{
swaap_select.contact(chr_input = NULL)
}
\arguments{
\item{chr_input}{An optional character
vector, additional columns to include.}
}
\value{
A character vector.
}
\description{
Function to select standard contact
information columns.
}
