% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Select_variables.R
\name{swaap_select.demographics}
\alias{swaap_select.demographics}
\title{Select Demographic Measures}
\usage{
swaap_select.demographics(chr_input = NULL)
}
\arguments{
\item{chr_input}{An optional character
vector, additional columns to include.}
}
\value{
A character vector.
}
\description{
Function to select standard demographic
measure columns.
}
\author{
Kevin Potter
}
