% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Select_variables.R
\name{swaap_select.experience}
\alias{swaap_select.experience}
\title{Select School Experience Items}
\usage{
swaap_select.experience(chr_input = NULL)
}
\arguments{
\item{chr_input}{An optional character
vector, additional columns to include.}
}
\value{
A character vector.
}
\description{
Function to select columns for school
experience items.
}
\author{
Kevin Potter
}
