% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Select_variables.R
\name{swaap_select.intermittent}
\alias{swaap_select.intermittent}
\title{Select Intermittent Variables}
\usage{
swaap_select.intermittent(chr_input = NULL)
}
\arguments{
\item{chr_input}{An optional character
vector, additional columns to include.}
}
\value{
A character vector.
}
\description{
Function to select columns for variables that
were measured at intermittent time points.
}
\author{
Kevin Potter
}
