% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Select_variables.R
\name{swaap_select.inventories}
\alias{swaap_select.inventories}
\title{Select Inventories}
\usage{
swaap_select.inventories(
  chr_input = NULL,
  chr_measures = c("AUDIT", "PHQ4", "ERS", "ADDI"),
  lgc_items = FALSE
)
}
\arguments{
\item{chr_input}{An optional character
vector, additional columns to include.}

\item{chr_measures}{A character vector, the
different measures to select, including
\code{'AUDIT'}, \code{'PHQ4'}, \code{'ERS'},
and \code{'ADDI'}.}

\item{lgc_items}{A logical value; if \code{TRUE}
includes the individual items for each measure,
otherwise includes total scores and subscale
scores.}
}
\value{
A character vector.
}
\description{
Function to select columns for inventories.
}
\author{
Kevin Potter
}
