% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Select_variables.R
\name{swaap_select.linked}
\alias{swaap_select.linked}
\title{Select Details on Linkage}
\usage{
swaap_select.linked(chr_input = NULL, lgc_subset = TRUE)
}
\arguments{
\item{chr_input}{An optional character
vector, additional columns to include.}

\item{lgc_subset}{A logical value; if
\code{TRUE} returns a subset of the
columns, the most useful typical details.}
}
\value{
A character vector.
}
\description{
Function to select columns detailing
linkage performance.
}
\author{
Kevin Potter
}
