% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Select_variables.R
\name{swaap_select.linking}
\alias{swaap_select.linking}
\title{Select Linking Items}
\usage{
swaap_select.linking(
  chr_input = NULL,
  lgc_district = FALSE,
  lgc_original = FALSE,
  lgc_fastLink = FALSE,
  lgc_all = FALSE
)
}
\arguments{
\item{chr_input}{An optional character
vector, additional columns to include.}

\item{lgc_district}{A logical value; if
\code{TRUE} uses district codes instead
of school codes.}

\item{lgc_original}{A logical value; if
\code{TRUE} uses original variable
names instead of recoded ones.}

\item{lgc_fastLink}{A logical value; if
\code{TRUE} returns column names for
variables specifically formatted to
work with \link[fastLink:fastLink]{fastLink::fastLink}.}

\item{lgc_all}{A logical value; if
\code{TRUE} returns all possible
linking items.}
}
\value{
A character vector.
}
\description{
Function to select standard linking
item columns.
}
\author{
Kevin Potter
}
