% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Select_variables.R
\name{swaap_select.misc}
\alias{swaap_select.misc}
\title{Select Variables for Miscellaneous Items}
\usage{
swaap_select.misc(chr_input = NULL)
}
\arguments{
\item{chr_input}{An optional character
vector, additional columns to include.}
}
\value{
A character vector.
}
\description{
Function to select variables for
miscellaneous items.
}
\author{
Kevin Potter
}
