% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Select_variables.R
\name{swaap_select.quality}
\alias{swaap_select.quality}
\title{Select Variables for Quality Control}
\usage{
swaap_select.quality(chr_input = NULL)
}
\arguments{
\item{chr_input}{An optional character
vector, additional columns to include.}
}
\value{
A character vector.
}
\description{
Function to select variables with
quality control details.
}
