% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Select_variables.R
\name{swaap_select}
\alias{swaap_select}
\title{Select Columns}
\usage{
swaap_select(dtf_data, chr_columns)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}

\item{chr_columns}{A character vector, the
columns in \code{dtf_data} to select
(columns that do not exist are ignored).}
}
\value{
A data frame.
}
\description{
Function to select specified columns
in a data set.
}
\author{
Kevin Potter
}
