% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Select_variables.R
\name{swaap_select.SBIRT}
\alias{swaap_select.SBIRT}
\title{Select Variables Denoting SBIRT Sample}
\usage{
swaap_select.SBIRT(chr_input = NULL)
}
\arguments{
\item{chr_input}{An optional character
vector, additional columns to include.}
}
\value{
A character vector.
}
\description{
Function to select variables tracking
SBIRT sample details.
}
