% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Select_variables.R
\name{swaap_select.substances}
\alias{swaap_select.substances}
\title{Select Variables Measuring Substance Use}
\usage{
swaap_select.substances(chr_input = NULL, lgc_SBIRT = FALSE)
}
\arguments{
\item{chr_input}{An optional character
vector, additional columns to include.}

\item{lgc_SBIRT}{A logical value; if \code{TRUE}
includes additional measures included for
schools and grades undergoing SBIRT.}
}
\value{
A character vector.
}
\description{
Function to select variables with self-reported
substance use details.
}
\author{
Kevin Potter
}
