% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Select_variables.R
\name{swaap_select.suicidality}
\alias{swaap_select.suicidality}
\title{Select Variables on Suicidality}
\usage{
swaap_select.suicidality(chr_input = NULL)
}
\arguments{
\item{chr_input}{An optional character
vector, additional columns to include.}
}
\value{
A character vector.
}
\description{
Function to select variables on suicidality.
}
\author{
Kevin Potter
}
