% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R06-Simulate_functions.R
\name{swaap_simulate}
\alias{swaap_simulate}
\title{Simulate data}
\usage{
swaap_simulate(chr_type, chr_subtype = "", ...)
}
\arguments{
\item{chr_type}{A character string, the type
of data to simulate.}

\item{chr_subtype}{A character string, the sub-type
of data to simulate.}

\item{...}{Additional arguments to pass to the
internal data generation functions.}
}
\value{
A data frame.
}
\description{
Function to simulate assorted data relevant to
the school-wide assessment project.
}
\examples{
dtf_debug <- swaap_simulate( 'link' )

}
\author{
Kevin Potter
}
