% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R02-Add_variables.R
\name{swaap_add.substances}
\alias{swaap_add.substances}
\title{Add Substance Use Variables}
\usage{
swaap_add.substances(dtf_data, chr_substance)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}

\item{chr_substance}{A character string,
the substance as originally labeled (either
\code{'Alcohol'}, \code{'Cannabis'},
\code{'Vapes'}, \code{'Cigarettes'},
\code{'Cigars'}, \code{'Smokeless'}}
}
\value{
A data frame with additional
columns for lifetime use, integer
and character versions of the 7-point
past-month use rating, number of days
used, and past-year use. Note that
only a subset of grades and time points
would have been asked questions for
number of days used and past-year use.
}
\description{
Function to add substance use variables
along with quality control checks.
}
\author{
Kevin Potter
}
