% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Data_functions.R
\name{swaap_data.download}
\alias{swaap_data.download}
\title{Download School-wide assessment Data}
\usage{
swaap_data.download(
  chr_dropbox,
  chr_data = "2023 Fall",
  lgc_complete = FALSE,
  lgc_rename = TRUE,
  chr_copy_to = NULL,
  lgc_silent = FALSE
)
}
\arguments{
\item{chr_dropbox}{A character string, the local
path to the user's Dropbox folder.}

\item{chr_data}{A character vector, the data sets
to copy, in the format \code{'<Year> <Semester>'}
(e.g., \code{c( '2023 Fall', '2024 Fall' )}).}

\item{lgc_complete}{A logical value; if \code{TRUE}
downloads the full data set that includes
confidential patient health information
(only set to \code{TRUE} if you are positive you
have full access to the data).}

\item{lgc_rename}{A logical value; if \code{TRUE}
renames source files to the format:
\code{'SWA-Surveys-<Year><Semester>-<Date>.csv'}}

\item{chr_copy_to}{A character string, the folder
to which data should be copied. If \code{NULL}
files will not be copied.}

\item{lgc_silent}{A logical value; if \code{TRUE}
suppresses warning messages.}
}
\value{
A data frame with the file paths (both full
and truncated to within the CAM Data Dropbox folder)
for the requested data sets. If \code{chr_copy_to}
is not \code{NULL} as a side effect copies the files
from the Dropbox folder to the user-specified local
folder.
}
\description{
Function to copy school-wide assessment
data from Dropbox to a local folder, or
at least provide path to specified data set.
}
\author{
Kevin Potter
}
