% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Linking_code.R
\name{swaap_link}
\alias{swaap_link}
\title{Link Records Across Time Points}
\usage{
swaap_link(
  dtf_long,
  chr_method = "dissimilarity",
  lgc_remove_duplicates = FALSE,
  lst_sets = NULL,
  obj_items = NULL,
  lst_combos = NULL,
  lst_missing = NULL,
  fun_trim_duplicates = NULL,
  chr_progress = "bar"
)
}
\arguments{
\item{dtf_long}{A data frame, assumed to be standard processed
school-wide assessment data with the columns
\code{'SSS.INT.TimePoint'} and \code{'SSS.INT.LongitudinalWave'}
as well as the linking code items.}

\item{chr_method}{A character string, either
\code{'dissimilarity'} (to link using dissimilarity scores)
or \code{'group_by'} (to link using simpler method based
on \link[dplyr:group_by]{dplyr::group_by}).}

\item{lgc_remove_duplicates}{A logical value; if \code{TRUE}
function attempts to identify and trim duplicate records
within a time point.}

\item{lst_sets}{A list of lists, with each sublist specifying
\code{'Base'} and \code{'Add'} logical vectors for the pair of data
subsets in \code{dtf_long} to link over (e.g., \code{'Base'} would
subset the first time point and \code{'Add'} would subset the second
time point).}

\item{obj_items}{Either a character vector with the column
names for the linking items, or a list of character vectors,
one vector for each set defined in \code{lst_sets}.
Passing a list with separate vectors allows using different
linking items for different sets when necessary. If
\code{NULL} attempts to identify linking items based on
standardized names.}

\item{lst_combos}{A list of lists, where each sublist consists of
an integer vector indexing the combination of linking items to
consider in order of priority. One sublist of integer vectors must
be defined for each set defined by \code{lst_sets}. For a
given sublist, indices apply to the character vector defined
for the relevant set in \code{obj_items}, meaning that if
character vectors differ across sets, indices should be defined
accordingly.}

\item{lst_missing}{A list of lists, where each sublist consists of
an integer vector indicating which items should be checked for
missingness when linking using a given combo (therefore
\code{lst_missing} should match in structure to \code{lst_combos}).
When specifying additional items beyond those listing in the given
combo, ensures that if these extra items are non-missing for a
record, the record will not be linked. This behavior can be
suppressed by supplying \code{c()} instead of an integer vector,
which in turn allows for matches with a dissimilarity score
greater than 0.}

\item{fun_trim_duplicates}{An optional function that returns
a logical vector for the subset of duplicate records
indicating which records should be kept (return \code{TRUE}).
Default is to select record with the highest completion rate.}

\item{chr_progress}{A character string, either \code{'bar'}
(progress bar to track function completion), \code{'section'}
(lists completed sections to track function completion), or
\code{''} to not track progress.}
}
\value{
A data frame.
}
\description{
Function to link records (e.g., across different time points)
using a set of linking items.
}
\examples{
# Generate demonstration data
dtf_long <- swaap_data( c( 'link', 'demo' ) )
# Record linkage
dtf_linked <- swaap_link( dtf_long )

}
\author{
Michael Pascale; Kevin Potter
}
