% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Linking_code.R
\name{swaap_link.report.by_ID}
\alias{swaap_link.report.by_ID}
\title{Linkage Patterns by Linked Identifier}
\usage{
swaap_link.report.by_ID(dtf_linked, lgc_update = FALSE)
}
\arguments{
\item{dtf_linked}{A data frame, output from the
\link{swaap_link} function. Must have the columns
\code{IDN.CHR.Linked.ID} and \code{SSS.INT.TimePoint}.}
}
\value{
A wide-form data frame with one row
per linked ID along with the pattern of linked
time points (e.g., \code{'0-1'} means a link
from baseline to the first time point).
}
\description{
Function to determine the pattern of linked time
points for each linked ID.
}
\author{
Kevin Potter
}
