% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-Recode_variables.R
\name{swaap_recode.experience}
\alias{swaap_recode.experience}
\title{Recode School Experience Items}
\usage{
swaap_recode.experience(dtf_data)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}
}
\value{
A data frame with the additional variables
\code{'SBJ.LGC.Experience.PlaySports'},
\code{'SBJ.LGC.Experience.SuspensionsAny'},
\code{'SBJ.LGC.Experience.SuspensionsDrug'},
\code{'SBJ.CHR.Experience.GradesInSchool'}, and
\code{'SBJ.INT.Experience.GradesInSchool'}.
}
\description{
Recodes school experience items for easy selection.
}
\author{
Kevin Potter
}
