% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-Recode_variables.R
\name{swaap_recode.inventories}
\alias{swaap_recode.inventories}
\title{Recode Inventory Variables}
\usage{
swaap_recode.inventories(dtf_data)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}
}
\value{
A data frame with the additional variables
\code{'INV.INT.AUDIT.Total'},
\code{'INV.INT.AUDIT.Q1'} to
\code{'INV.INT.AUDIT.Q10'},
\code{'INV.INT.ADDI.Total'},
\code{'INV.INT.ADDI.Q1'} to
\code{'INV.INT.ADDI.Q15'}.
}
\description{
Renames inventory variables for easy selection.
}
\author{
Kevin Potter
}
