% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-Recode_variables.R
\name{swaap_recode.linking}
\alias{swaap_recode.linking}
\title{Recode Linking Items}
\usage{
swaap_recode.linking(dtf_data)
}
\arguments{
\item{dtf_data}{A data frame, assumed to
follow the standardized format for the
school-wide assessment data.}
}
\value{
A data frame with the additional variables
\code{'SBJ.INT.Link.SchoolCode'},
\code{'SBJ.INT.Link.SchoolID'},
\code{'SBJ.CHR.Link.Sex'},
\code{'SBJ.CHR.Link.BirthYearMonth'},
\code{'SBJ.CHR.Link.OlderSiblings'},
\code{'SBJ.CHR.Link.EyeColor'}, and
\code{'SBJ.CHR.Link.MiddleInitial'}.
}
\description{
Renames linking items for easy selection.
}
\author{
Kevin Potter
}
